core: New schema methods on terraform.ResourceProvider
In order to parse provider, resource and data source configuration from
HCL2 config files, we need to know the relevant configuration schema.
These new methods allow Terraform Core to request these from a provider.

This is a breaking change to this interface, so all of its implementers
in this package are updated too. This includes concrete implementations
of the new methods in helper/schema that use the schema conversion code
added in an earlier commit to produce a configschema.Block automatically.

Plugins compiled against prior versions of helper/schema will not have
support for these methods, and so calls to them will fail. Callers of
these new methods will therefore need to be resilient to such failures
and proceed under the assumption that the provider does not yet have
schema support. This careful handling will need to persist until next time
we increment the plugin protocol version, at which point we can make
the breaking change of requiring this information to be available.